Move #presentModal() into correct class (deferrableInteraction)
